requirements.txt
django
djangorestframework
django-cors-headers
django-debug-toolbar
django-rest-authemail
djangorestframework-jwt
psycopg2
python-decouple
django-webpack-loader
gunicorn
Whitenoise